export base crate to core, and create re-exporting crate
