+++ bevy_script_api/Cargo.toml
@@ -17,8 +17,8 @@ rhai = ["bevy_mod_scripting_rhai"]
 
 [dependencies]
 bevy = { version = "0.8.0", default-features = false, features=["bevy_asset","bevy_gltf","bevy_animation","bevy_core_pipeline","bevy_ui","bevy_pbr","bevy_render","bevy_text","bevy_sprite","filesystem_watcher"]}
-bevy_mod_scripting_derive={path="../bevy_mod_scripting_derive",version="0.1.1"}
-bevy_mod_scripting={path="../",version="0.1.1"}
+bevy_mod_scripting_derive = { path="../bevy_mod_scripting_derive", version = "0.1.1"}
+bevy_mod_scripting_core = { path="../bevy_mod_scripting_core", version = "0.1.1"}
 parking_lot="0.12.1"
 paste="1.0.7"
 thiserror="1.0.32"
