+++ languages/bevy_mod_scripting_lua/Cargo.toml
@@ -23,7 +23,6 @@ lua53 = ["tealr/mlua_lua53"]
 lua54 = ["tealr/mlua_lua54"]
 luajit = ["tealr/mlua_luajit"]
 luajit52 = ["tealr/mlua_luajit52"]
-
 mlua_serialize = ["tealr/mlua_serialize"]
 mlua_macros = ["tealr/mlua_macros"]
 
@@ -33,7 +32,7 @@ path="src/lib.rs"
 
 [dependencies]
 bevy= { version = "0.8.0", default-features = false}
-bevy_mod_scripting = {path="../../", version="0.1.1"}
+bevy_mod_scripting_core = {path="../../bevy_mod_scripting_core", version="0.1.1"}
 tealr = { version = "0.9.0-alpha1", features=["mlua_vendored","mlua_send"]}
 parking_lot = "0.12.1"
 serde_json = "1.0.81"
